Fix!: render fns should return nil if no content.
The following functions have been updated and tested to make sure that
they return nil if they don't have any return values
- (render :related-files)
- (render :related-files)

This is marked as a breaking change only in that it may inform how you
are using layouts. This change is more useful for using `when-let`
the value to conditionally render content if it exists.

Other work:

Docs: fix incorrect link to folding document.

Docs: Document nil return val on related-files.

Test: Add test for backlinks expecting nil.

Test: use ul in render-firn-file-tags

Test: tag functions.